% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nan.R
\name{DNA_to_prt}
\alias{DNA_to_prt}
\title{DNA sequences were coded as protein sequences}
\usage{
DNA_to_prt(DNASEQ = NULL, splits = F)
}
\arguments{
\item{DNASEQ}{DNA sequences}

\item{splits}{Whether to divide}
}
\value{

}
\description{
DNA sequences were coded as protein sequences
}
