% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nan.R
\name{DNA_to_split}
\alias{DNA_to_split}
\title{Title Codon splitting of DNA sequences}
\usage{
DNA_to_split(DNASEQ = NULL)
}
\arguments{
\item{DNASEQ}{DNA sequences}
}
\value{

}
\description{
Title Codon splitting of DNA sequences
}
